Don't delay initial handshakes right after creating a peer.
Initializing a peer would default the lastSentHandshake
time to ~1 second in the future, preventing any initial handshake from
triggering until a retry period had passed. This seems
counterproductive, so I removed that. Now we connect extremely fast
after a reconfig.

But I don't know why this was here, so maybe it's risky to remove.