[Tailscale] HACKS: wintun: faster discovery, skip interface create/delete.
Remove wintun "network name" check. We don't need this for tailscale
because we only need one network interface total. The end user might
rename the network interface, so this is not a safe string to use for
finding interfaces anyway.

Don't ever create new wintun interfaces. If we can't find the default
one, just fail. Creating interfaces requires admin access, which we
don't want to be necessary.