device: Simple packet filtering support.
An app can inject a FilterPacket() function that can decide whether to
accept or reject a given packet based on its contents.  The default is
the old behaviour.

It runs right after decrypt, in the same goroutine, in order to take
advantage of the same parallelism. Beware that this means your
FilterPacket() function might be run many times in parallel (with
different buffers), so beware that it needs to be reentrant.

TODO: this adds a device.SetFilterPacket() function.
  I want to be able to update the packet filter without restarting
  wireguard. This is a bit ugly because it introduces a lock into a
  high-throughput pathway, which means the final version of this feature
  probably can't really work like this.