[Tailscale] shrink queue size on iOS
This is a huge queue shrink, from 1024 to 64.
The effects on throughput aren't known yet because I'm on a weak
internet connection and don't have enough gear around to test it.
(This will change if I can get the tailscale app working well on
my macOS laptop.)

In conjunction with the zstd changes, this gets the latest iOS
app to work for (at least) 30 minutes without OOM.