[Tailscale] device: pass around *net.UDPAddr in addition to Endpoint.
An Endpoint represents the address we *think* a peer will use to talk
to us, but a packet may have arrived from a different source, so we
need to be able to represent them separately in a few places
(especially log messages).

On the other hand, when sending packets back to a given peer, we want
to use its primary endpoint, so we can't pass around only the UDPAddr,
we have to keep the old endpoint object as well.

To allow for more flexible Endpoint implementations, instead of
completely replacing the Endpoint object after validating a handshake,
add an UpdateDst() function instead.

[Based on a patch by David Crawshaw]