[tailscale] device: fix two data races with "Send handshake response" debug printing.
One race happens in production code, due to concurrent pointer updates.

The other only happens when using wireguard-go's own endpoint implementation,
which only happens in wireguard-go tests for our fork. For that one I
hacked in an ugly fix to remove the data race noise. A correct fix involves
upstreaming our additional debugging, and resolving wireguard-go's locking
and concurrency challenges at that point.

DO NOT UPSTREAM THIS. It's a set of fixes that only matters because other
code in this fork introduced a data race. The correct way to deal with that
is to make the original change safe and upstreamable. This is just a
temporary fix.

Passes `go test -count=1000 -race ./device`.